fixed connect hang when max_udp_payload_size overflow

+++ include/quicly.h
@@ -267,7 +267,7 @@ typedef struct st_quicly_transport_parameters_t {
     /**
      *
      */
-    uint64_t max_udp_payload_size;
+    uint16_t max_udp_payload_size;
     /**
      * quicly ignores the value set for quicly_context_t::transport_parameters
      */
